src/python/requirements.txt
openai==0.27.6
openai[embeddings]
jupyter==1.0.0
python-dotenv==1.0.0
matplotlib==3.7.1
panel==0.14.4
tiktoken
mwclient
mwparserfromhell
PyPDF2
openai[embeddings]
tiktoken==0.4.0
langchain==0.0.187
atlassian-python-api==3.38.0
fastapi==0.95.2
websockets==11.0.3
uvicorn[standard]==0.22.0
pinecone-client==2.2.1
pypdf